+++ docs/src/whats-new.md
@@ -7,6 +7,16 @@ description: Learn about the latest features and improvements in LogLayer
 
 - [`loglayer` Changelog](/core-changelogs/loglayer-changelog)
 
+## Feb 11, 2025
+
+Major version 6 is out! 🎉
+
+This adds a new feature called [Context Managers](/context-managers/).
+
+### Breaking Changes
+
+- The `linkParentContext` configuration option has been removed. See the [migration guide](/migrating) for more information on how to replicate the functionality.
+
 ## Feb 8, 2025
 
 - Added `linkParentContext` configuration option to allow child loggers to link to their parent's context. This enables changes to the context in either logger to affect both parent and child. [Learn more](./logging-api/child-loggers.html#shared-context-reference)